# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Instagram-clone API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get currently authenticated user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /users/{username}:
    get:
      operationId: getUserProfile
      tags: [users]
      summary: Get user profile by username
      parameters:
        - name: username
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /users/{username}/posts:
    get:
      operationId: getUserPosts
      tags: [users]
      summary: Get posts by user
      parameters:
        - name: username
          in: path
          required: true
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 12
      responses:
        "200":
          description: User posts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PostsPage"
  /users/me/profile:
    patch:
      operationId: updateMyProfile
      tags: [users]
      summary: Update current user's profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProfileUpdate"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
  /users/me/avatar:
    post:
      operationId: uploadAvatar
      tags: [users]
      summary: Upload profile picture (base64)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MediaUploadInput"
      responses:
        "200":
          description: Updated avatar URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UploadMediaResponse"
  /users/{username}/follow:
    post:
      operationId: followUser
      tags: [users]
      summary: Follow a user
      parameters:
        - name: username
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Followed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FollowResponse"
  /users/{username}/unfollow:
    post:
      operationId: unfollowUser
      tags: [users]
      summary: Unfollow a user
      parameters:
        - name: username
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Unfollowed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FollowResponse"
  /users/{username}/followers:
    get:
      operationId: getFollowers
      tags: [users]
      summary: Get followers of a user
      parameters:
        - name: username
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of followers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserSummary"
  /users/{username}/following:
    get:
      operationId: getFollowing
      tags: [users]
      summary: Get users that this user follows
      parameters:
        - name: username
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of following
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserSummary"
...
  /vault/pin:
    post:
      operationId: setVaultPin
      tags: [vault]
      summary: Set or update the vault master PIN
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VaultPinInput"
      responses:
        "200":
          description: PIN set successfully
  /vault/unlock:
    post:
      operationId: unlockVault
      tags: [vault]
      summary: Verify vault PIN and get vaulted conversations
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VaultUnlockInput"
      responses:
        "200":
          description: Vault unlocked with conversation list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/VaultedConversation"
        "401":
          description: Wrong PIN
  /vault/add:
    post:
      operationId: addToVault
      tags: [vault]
      summary: Add a conversation to the vault
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VaultPostInput"
      responses:
        "200":
          description: Added to vault
  /vault/{conversationId}:
    delete:
      operationId: removeFromVault
      tags: [vault]
      summary: Remove a conversation from the vault
      parameters:
        - name: conversationId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Removed from vault
  /vault/{conversationId}/unlock:
    post:
      operationId: unlockVaultConversation
      tags: [vault]
      summary: Verify a conversation PIN inside the vault
      parameters:
        - name: conversationId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VaultConvUnlockInput"
      responses:
        "200":
          description: Conversation unlocked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VaultedConversation"
        "401":
          description: Wrong PIN
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    # ── Auth ──────────────────────────────────────────────────────────────
...
    UserProfile:
      type: object
      properties:
        id:
          type: string
        username:
          type: string
        fullName:
          type: string
        bio:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        postsCount:
          type: integer
        followersCount:
          type: integer
        followingCount:
          type: integer
        isFollowing:
          type: boolean
        isMe:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - username
        - fullName
        - postsCount
        - followersCount
        - followingCount
        - isFollowing
        - isMe
        - createdAt
    UserSummary:
      type: object
      properties:
        id:
          type: string
        username:
          type: string
        fullName:
          type: string
        avatarUrl:
          type: ["string", "null"]
        isFollowing:
          type: boolean
      required:
        - id
        - username
        - fullName
        - isFollowing
    ProfileUpdate:
      type: object
      properties:
        fullName:
          type: string
        bio:
          type: string
        username:
          type: string
    FollowResponse:
      type: object
      properties:
        isFollowing:
          type: boolean
        followersCount:
          type: integer
      required:
        - isFollowing
        - followersCount

    # ── Posts ──────────────────────────────────────────────────────────────
    Post:
      type: object
      properties:
        id:
          type: string
        caption:
          type: ["string", "null"]
        mediaUrl:
          type: string
        mediaType:
          type: string
          enum: [image, video]
        audience:
          type: string
          enum: [everyone, close_friends]
        location:
          type: ["string", "null"]
        altText:
          type: ["string", "null"]
        commentsDisabled:
          type: boolean
        additionalMediaUrls:
          type: array
          items:
            type: string
        likesCount:
          type: integer
        commentsCount:
          type: integer
        isLiked:
          type: boolean
        isSaved:
          type: boolean
        author:
          $ref: "#/components/schemas/UserSummary"
        createdAt:
          type: string
      required:
        - id
        - mediaUrl
        - mediaType
        - audience
        - commentsDisabled
        - additionalMediaUrls
        - likesCount
        - commentsCount
        - isLiked
        - isSaved
        - author
        - createdAt
    PostsPage:
      type: object
      properties:
        posts:
          type: array
          items:
            $ref: "#/components/schemas/Post"
        hasMore:
          type: boolean
        total:
          type: integer
      required:
        - posts
        - hasMore
        - total
...
    MediaUploadInput:
      type: object
      properties:
        data:
          type: string
          description: Base64 encoded file data
        mimeType:
          type: string
      required:
        - data
        - mimeType
    UploadMediaResponse:
      type: object
      properties:
        url:
          type: string
        publicId:
          type: string
        resourceType:
          type: string
      required:
        - url
        - publicId
        - resourceType
...
    VaultPinInput:
      type: object
      properties:
        pin:
          type: string
        currentPin:
          type: string
      required:
        - pin
    VaultUnlockInput:
      type: object
      properties:
        pin:
          type: string
      required:
        - pin
    VaultPostInput:
      type: object
      properties:
        conversationId:
          type: string
        pin:
          type: string
      required:
        - conversationId
        - pin
    VaultConvUnlockInput:
      type: object
      properties:
        pin:
          type: string
      required:
        - pin
    VaultedConversation:
      type: object
      properties:
        conversationId:
          type: string
        otherUser:
          $ref: "#/components/schemas/UserSummary"
        lastMessage:
          type: ["string", "null"]
        lastMessageAt:
          type: ["string", "null"]
        unreadCount:
          type: integer
      required:
        - conversationId
        - otherUser
        - unreadCount